+++ modules/tensor_mechanics/test/tests/ad_anisotropic_creep/ad_aniso_creep_z_3d.i
@@ -28,7 +28,6 @@
   []
 []
 
-
 [GlobalParams]
   displacements = 'disp_x disp_y disp_z'
   volumetric_locking_correction = true
